+++ AppLovin/com/mopub/mobileads/AppLovinInterstitial.java
@@ -15,9 +15,12 @@ import com.applovin.sdk.AppLovinAdLoadListener;
 import com.applovin.sdk.AppLovinAdSize;
 import com.applovin.sdk.AppLovinAdVideoPlaybackListener;
 import com.applovin.sdk.AppLovinErrorCodes;
+import com.applovin.sdk.AppLovinPrivacySettings;
 import com.applovin.sdk.AppLovinSdk;
 import com.applovin.sdk.AppLovinSdkSettings;
+import com.mopub.common.MoPub;
 import com.mopub.common.logging.MoPubLog;
+import com.mopub.common.privacy.PersonalInfoManager;
 
 import java.util.HashMap;
 import java.util.LinkedList;
@@ -47,6 +50,15 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
 
     @Override
     public void loadInterstitial(final Context context, final CustomEventInterstitialListener listener, final Map<String, Object> localExtras, final Map<String, String> serverExtras) {
+
+        PersonalInfoManager personalInfoManager = MoPub.getPersonalInformationManager();
+
+        // Pass the user consent from the MoPub SDK to AppLovin as per GDPR
+        if (personalInfoManager != null) {
+            boolean canCollectPersonalInfo = personalInfoManager.canCollectPersonalInformation();
+            AppLovinPrivacySettings.setHasUserConsent(canCollectPersonalInfo, context);
+        }
+
         MoPubLog.d("Requesting AppLovin interstitial with serverExtras: " + serverExtras + " and localExtras: " + localExtras);
 
         // SDK versions BELOW 7.2.0 require a instance of an Activity to be passed in as the context
@@ -81,7 +93,7 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
             // Otherwise, use the Zones API
             else {
                 // Dynamically load an ad for a given zone without breaking backwards compatibility for publishers on older SDKs
-                sdk.getAdService().loadNextAdForZoneId( zoneId, this );
+                sdk.getAdService().loadNextAdForZoneId(zoneId, this);
             }
         }
     }
@@ -90,7 +102,7 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
     public void showInterstitial() {
         final AppLovinAd preloadedAd = dequeueAd(zoneId);
         if (preloadedAd != null) {
-            final AppLovinInterstitialAdDialog interstitialAd = AppLovinInterstitialAd.create( sdk, context );
+            final AppLovinInterstitialAdDialog interstitialAd = AppLovinInterstitialAd.create(sdk, context);
             interstitialAd.setAdDisplayListener(this);
             interstitialAd.setAdClickListener(this);
             interstitialAd.setAdVideoPlaybackListener(this);