+++ OnebyAOL/com/mopub/mobileads/MillennialInterstitial.java
@@ -15,8 +15,12 @@ import com.millennialmedia.MMSDK;
 import com.millennialmedia.internal.ActivityListenerManager;
 import com.mopub.common.MoPub;
 import com.mopub.common.logging.MoPubLog;
+import com.mopub.common.privacy.PersonalInfoManager;
 
 import java.util.Map;
+
+import static com.millennialmedia.MMSDK.setConsentData;
+import static com.millennialmedia.MMSDK.setConsentRequired;
 
 /**
  * Compatible with version 6.6 of the Millennial Media SDK.
@@ -53,6 +57,21 @@ final class MillennialInterstitial extends CustomEventInterstitial {
         if (context instanceof Activity) {
             try {
                 MMSDK.initialize((Activity) context, ActivityListenerManager.LifecycleState.RESUMED);
+
+                PersonalInfoManager personalInfoManager = MoPub.getPersonalInformationManager();
+
+                if (personalInfoManager != null) {
+                    boolean gdprApplies = personalInfoManager.gdprApplies();
+                    boolean canCollectPersonalInfo = personalInfoManager.canCollectPersonalInformation();
+
+                    // Set if GDPR applies / if consent is required
+                    setConsentRequired(gdprApplies);
+
+                    // Pass the user consent from the MoPub SDK to One by AOL as per GDPR
+                    if (canCollectPersonalInfo) {
+                        setConsentData("mopub", "1");
+                    }
+                }
             } catch (IllegalStateException e) {
                 MoPubLog.d("Exception occurred initializing the MM SDK.", e);
                 interstitialListener.onInterstitialFailed(MoPubErrorCode.INTERNAL_ERROR);