+++ AppLovin/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Changelog
-  
+  * 8.0.1.0
+    * This version of the adapters has been certified with AppLovin 8.0.1.
+    * Pass user consent collected from the MoPub SDK to the AppLovin SDK per GDPR.
   * 7.8.6.3
     * Removed reflection from all adapters, as there are cases on MultiDexed applications, where reflection may not be able to find the target class if it's in another dex file. This resolves this issue: https://github.com/mopub/mopub-android-mediation/issues/19#issuecomment-382488061.
   * 7.8.6.2