+++ Flurry/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+  * 10.1.0.0
+    * This version of the adapters has been certified with Flurry 10.1.0.
+    * Pass user consent collected from the MoPub SDK to the Flurry SDK per GDPR.
+
   * 9.0.0.0
     * This version of the adapters has been certified with Flurry 9.0.0.