+++ IronSource/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+  * 6.7.9.0
+    * This version of the adapters has been certified with IronSource 6.7.9.
+    * Pass user consent collected from the MoPub SDK to the ironSource SDK per GDPR.
+
   * 6.7.8.0
     * This version of the adapters has been certified with IronSource 6.7.8.