+++ Vungle/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+  * 6.2.5.0
+    * This version of the adapters has been certified with Vungle 6.2.5.
+    * Pass user consent collected from the MoPub SDK to the Vungle SDK per GDPR.
+
   * 5.3.2.0
     * This version of the adapters has been certified with Vungle 5.3.2.